+++ crates/client/commitment-state-diff/src/lib.rs
@@ -26,7 +26,6 @@ pub struct BlockDAData {
     pub state_diff: ThinStateDiff,
     pub num_addr_accessed: usize,
     pub block_number: u64,
-    pub config_hash: StarkHash,
     pub new_state_root: StarkHash,
     pub previous_state_root: StarkHash,
 }
@@ -76,7 +75,7 @@ where
                 Poll::Ready(Some(storage_notification)) => {
                     let block_hash = storage_notification.block;
 
-                    match build_commitment_state_diff::<B, C, H>(
+                    match build_commitment_state_diff::<B, C>(
                         self_as_mut.client.clone(),
                         self_as_mut.backend.clone(),
                         storage_notification,
@@ -137,7 +136,7 @@ enum BuildCommitmentStateDiffError {
     FailedToGetConfigHash(#[from] sp_api::ApiError),
 }
 
-fn build_commitment_state_diff<B: BlockT, C, H>(
+fn build_commitment_state_diff<B: BlockT, C>(
     client: Arc<C>,
     backend: Arc<mc_db::Backend<B>>,
     storage_notification: StorageNotification<B::Hash>,
@@ -146,7 +145,6 @@ where
     C: ProvideRuntimeApi<B>,
     C::Api: StarknetRuntimeApi<B>,
     C: HeaderBackend<B>,
-    H: HasherT,
 {
     let mut accessed_addrs: IndexSet<ContractAddress> = IndexSet::new();
     let mut commitment_state_diff = ThinStateDiff {
@@ -232,14 +230,11 @@ where
         mp_digest_log::find_starknet_block(digest)?
     };
 
-    let config_hash = client.runtime_api().config_hash(storage_notification.block)?;
-
     Ok(BlockDAData {
-        block_hash: current_block.header().hash::<H>().into(),
+        block_hash: current_block.header().hash().into(),
         state_diff: commitment_state_diff,
         num_addr_accessed: accessed_addrs.len(),
         block_number: current_block.header().block_number,
-        config_hash,
         // TODO: fix when we implement state root
         new_state_root: backend.temporary_global_state_root_getter(),
         previous_state_root: backend.temporary_global_state_root_getter(),