+++ crates/client/mapping-sync/src/sync_blocks.rs
@@ -1,6 +1,5 @@
 use mc_rpc_core::utils::get_block_by_block_hash;
 use mp_digest_log::{find_starknet_block, FindLogError};
-use mp_hashers::HasherT;
 use mp_transactions::get_transaction_hash;
 use num_traits::FromPrimitive;
 use pallet_starknet_runtime_api::StarknetRuntimeApi;
@@ -12,7 +11,7 @@ use sp_runtime::traits::{Block as BlockT, Header as HeaderT, Zero};
 
 use crate::block_metrics::BlockMetrics;
 
-fn sync_block<B: BlockT, C, BE, H>(
+fn sync_block<B: BlockT, C, BE>(
     client: &C,
     backend: &mc_db::Backend<B>,
     header: &B::Header,
@@ -23,7 +22,6 @@ where
     C: ProvideRuntimeApi<B>,
     C::Api: StarknetRuntimeApi<B>,
     BE: Backend<B>,
-    H: HasherT,
 {
     // Before storing the new block in the Madara backend database, we want to make sure that the
     // wrapped Starknet block it contains is the same that we can find in the storage at this height.
@@ -36,8 +34,8 @@ where
             let opt_storage_starknet_block = get_block_by_block_hash(client, substrate_block_hash);
             match opt_storage_starknet_block {
                 Ok(storage_starknet_block) => {
-                    let digest_starknet_block_hash = digest_starknet_block.header().hash::<H>();
-                    let storage_starknet_block_hash = storage_starknet_block.header().hash::<H>();
+                    let digest_starknet_block_hash = digest_starknet_block.header().hash();
+                    let storage_starknet_block_hash = storage_starknet_block.header().hash();
                     // Ensure the two blocks sources (chain storage and block digest) agree on the block content
                     if digest_starknet_block_hash != storage_starknet_block_hash {
                         Err(anyhow::anyhow!(
@@ -100,15 +98,10 @@ where
     }
 }
 
-fn sync_genesis_block<B: BlockT, C, H>(
-    _client: &C,
-    backend: &mc_db::Backend<B>,
-    header: &B::Header,
-) -> anyhow::Result<()>
+fn sync_genesis_block<B: BlockT, C>(_client: &C, backend: &mc_db::Backend<B>, header: &B::Header) -> anyhow::Result<()>
 where
     C: HeaderBackend<B>,
     B: BlockT,
-    H: HasherT,
 {
     let substrate_block_hash = header.hash();
 
@@ -119,7 +112,7 @@ where
         }
         Err(FindLogError::MultipleLogs) => return Err(anyhow::anyhow!("Multiple logs found")),
     };
-    let block_hash = block.header().hash::<H>();
+    let block_hash = block.header().hash();
     let mapping_commitment = mc_db::MappingCommitment::<B> {
         block_hash: substrate_block_hash,
         starknet_block_hash: block_hash.into(),
@@ -131,7 +124,7 @@ where
     Ok(())
 }
 
-fn sync_one_block<B: BlockT, C, BE, H>(
+fn sync_one_block<B: BlockT, C, BE>(
     client: &C,
     substrate_backend: &BE,
     madara_backend: &mc_db::Backend<B>,
@@ -143,7 +136,6 @@ where
     C::Api: StarknetRuntimeApi<B>,
     C: HeaderBackend<B> + StorageProvider<B, BE>,
     BE: Backend<B>,
-    H: HasherT,
 {
     let mut current_syncing_tips = madara_backend.meta().current_syncing_tips()?;
 
@@ -173,12 +165,12 @@ where
     };
 
     if operating_header.number() == &Zero::zero() {
-        sync_genesis_block::<_, _, H>(client, madara_backend, &operating_header)?;
+        sync_genesis_block::<_, _>(client, madara_backend, &operating_header)?;
 
         madara_backend.meta().write_current_syncing_tips(current_syncing_tips)?;
         Ok(true)
     } else {
-        sync_block::<_, _, _, H>(client, madara_backend, &operating_header, block_metrics)?;
+        sync_block::<_, _, _>(client, madara_backend, &operating_header, block_metrics)?;
 
         current_syncing_tips.push(*operating_header.parent_hash());
         madara_backend.meta().write_current_syncing_tips(current_syncing_tips)?;
@@ -186,7 +178,7 @@ where
     }
 }
 
-pub fn sync_blocks<B: BlockT, C, BE, H>(
+pub fn sync_blocks<B: BlockT, C, BE>(
     client: &C,
     substrate_backend: &BE,
     madara_backend: &mc_db::Backend<B>,
@@ -199,13 +191,12 @@ where
     C::Api: StarknetRuntimeApi<B>,
     C: HeaderBackend<B> + StorageProvider<B, BE>,
     BE: Backend<B>,
-    H: HasherT,
 {
     let mut synced_any = false;
 
     for _ in 0..limit {
         synced_any = synced_any
-            || sync_one_block::<_, _, _, H>(client, substrate_backend, madara_backend, sync_from, block_metrics)?;
+            || sync_one_block::<_, _, _>(client, substrate_backend, madara_backend, sync_from, block_metrics)?;
     }
 
     Ok(synced_any)