+++ crates/client/rpc/src/lib.rs
@@ -16,7 +16,7 @@ use std::marker::PhantomData;
 use std::sync::Arc;
 
 use blockifier::transaction::account_transaction::AccountTransaction;
-use blockifier::transaction::objects::ResourcesMapping;
+use blockifier::transaction::objects::{ResourcesMapping, TransactionExecutionInfo};
 use blockifier::transaction::transactions::L1HandlerTransaction;
 use errors::StarknetRpcApiError;
 use jsonrpsee::core::{async_trait, RpcResult};
@@ -28,6 +28,7 @@ pub use mc_rpc_core::{
     StarknetWriteRpcApiServer,
 };
 use mc_storage::OverrideHandle;
+use mp_block::BlockTransactions;
 use mp_felt::Felt252Wrapper;
 use mp_hashers::HasherT;
 use mp_simulations::SimulationFlags;
@@ -51,7 +52,6 @@ use sp_blockchain::HeaderBackend;
 use sp_core::H256;
 use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
 use sp_runtime::transaction_validity::InvalidTransaction;
-use sp_runtime::DispatchError;
 use starknet_api::block::BlockHash;
 use starknet_api::core::Nonce;
 use starknet_api::hash::StarkFelt;
@@ -160,7 +160,7 @@ where
             Ok(block) => block,
             Err(_) => return Err(StarknetRpcApiError::BlockNotFound),
         };
-        Ok(starknet_block.header().hash::<H>().into())
+        Ok(starknet_block.header().hash().into())
     }
 
     /// Returns the substrate block hash corresponding to the given Starknet block id
@@ -617,15 +617,12 @@ where
         })?;
 
         let calldata = Calldata(Arc::new(request.calldata.iter().map(|x| Felt252Wrapper::from(*x).into()).collect()));
-
         let result = self.do_call(
             substrate_block_hash,
             Felt252Wrapper(request.contract_address).into(),
             Felt252Wrapper(request.entry_point_selector).into(),
             calldata,
         )?;
-
-        let result = self.convert_error(substrate_block_hash, result)?;
 
         Ok(result.iter().map(|x| format!("{:#x}", x.0)).collect())
     }
@@ -741,13 +738,13 @@ where
                 if starting_block.is_ok() && current_block.is_ok() && highest_block.is_ok() {
                     // Convert block numbers and hashes to the respective type required by the `syncing` endpoint.
                     let starting_block_num = UniqueSaturatedInto::<u64>::unique_saturated_into(self.starting_block);
-                    let starting_block_hash = starting_block?.header().hash::<H>().0;
+                    let starting_block_hash = starting_block?.header().hash().0;
 
                     let current_block_num = UniqueSaturatedInto::<u64>::unique_saturated_into(best_number);
-                    let current_block_hash = current_block?.header().hash::<H>().0;
+                    let current_block_hash = current_block?.header().hash().0;
 
                     let highest_block_num = UniqueSaturatedInto::<u64>::unique_saturated_into(highest_number);
-                    let highest_block_hash = highest_block?.header().hash::<H>().0;
+                    let highest_block_hash = highest_block?.header().hash().0;
 
                     // Build the `SyncStatus` struct with the respective syn information
                     Ok(SyncStatusType::Syncing(SyncStatus {
@@ -834,7 +831,7 @@ where
 
         let starknet_block = get_block_by_block_hash(self.client.as_ref(), substrate_block_hash)?;
         let starknet_version = starknet_block.header().protocol_version;
-        let block_hash = starknet_block.header().hash::<H>();
+        let block_hash = starknet_block.header().hash();
 
         let transaction_hashes =
             starknet_block.transactions_hashes().map(|txh| Felt252Wrapper::from(*txh).into()).collect();
@@ -1095,7 +1092,7 @@ where
 
         let starknet_block = get_block_by_block_hash(self.client.as_ref(), substrate_block_hash)?;
 
-        let block_hash = starknet_block.header().hash::<H>();
+        let block_hash = starknet_block.header().hash();
         let starknet_version = starknet_block.header().protocol_version;
         let transactions = starknet_block.transactions().iter().map(|tx| to_starknet_core_tx(tx.clone())).collect();
 
@@ -1165,7 +1162,7 @@ where
             FieldElement::default()
         };
 
-        let starknet_block_hash = BlockHash(starknet_block.header().hash::<H>().into());
+        let starknet_block_hash = BlockHash(starknet_block.header().hash().into());
 
         let state_diff = self.get_state_diff(&starknet_block_hash).map_err(|e| {
             error!("Failed to get state diff. Starknet block hash: {starknet_block_hash}, error: {e}");
@@ -1173,7 +1170,7 @@ where
         })?;
 
         let state_update = StateUpdate {
-            block_hash: starknet_block.header().hash::<H>().into(),
+            block_hash: starknet_block.header().hash().into(),
             new_root: Felt252Wrapper::from(self.backend.temporary_global_state_root_getter()).into(),
             old_root,
             state_diff,
@@ -1378,7 +1375,7 @@ where
             transactions: transaction_hashes,
             // TODO: fill real prices
             l1_gas_price: ResourcePrice { price_in_fri: Default::default(), price_in_wei: Default::default() },
-            parent_hash: latest_block_header.hash::<H>().into(),
+            parent_hash: latest_block_header.hash().into(),
             sequencer_address: Felt252Wrapper::from(latest_block_header.sequencer_address).into(),
             starknet_version: latest_block_header.protocol_version.to_string(),
             timestamp: calculate_pending_block_timestamp(),
@@ -1399,7 +1396,7 @@ where
             transactions,
             // TODO: fill real prices
             l1_gas_price: ResourcePrice { price_in_fri: Default::default(), price_in_wei: Default::default() },
-            parent_hash: latest_block_header.hash::<H>().into(),
+            parent_hash: latest_block_header.hash().into(),
             sequencer_address: Felt252Wrapper::from(latest_block_header.sequencer_address).into(),
             starknet_version: latest_block_header.protocol_version.to_string(),
             timestamp: calculate_pending_block_timestamp(),
@@ -1431,7 +1428,7 @@ where
         let starknet_block: mp_block::Block = get_block_by_block_hash(self.client.as_ref(), substrate_block_hash)
             .map_err(|_e| StarknetRpcApiError::BlockNotFound)?;
         let block_header = starknet_block.header();
-        let block_hash = block_header.hash::<H>().into();
+        let block_hash = block_header.hash().into();
         let block_number = block_header.block_number;
 
         let transaction =
@@ -1483,9 +1480,6 @@ where
 
         let messages_sent = messages.into_iter().map(starknet_api_to_starknet_core_message_to_l1).collect();
 
-        // TODO: use re_execute_transaction in order to rebuild the correct state and not have to skip the
-        // validation phase. It will return more accurate fees
-        let skip_validate = true;
         let parent_block_hash = self
             .substrate_block_hash_from_starknet_block(BlockId::Hash(
                 Felt252Wrapper::from(block_header.parent_block_hash).into(),
@@ -1494,8 +1488,9 @@ where
                 error!("Parent Block not found: {e}");
                 StarknetRpcApiError::BlockNotFound
             })?;
-        let simulation = self.simulate_tx(parent_block_hash, transaction.clone(), skip_validate, fee_disabled)?;
-        let execution_resources = actual_resources_to_execution_resources(simulation.actual_resources);
+        let execution_info =
+            self.get_transaction_execution_info(parent_block_hash, starknet_block.transactions(), transaction_hash)?;
+        let execution_resources = actual_resources_to_execution_resources(execution_info.actual_resources);
         let transaction_hash = Felt252Wrapper::from(transaction_hash).into();
 
         let receipt = match transaction {
@@ -1563,6 +1558,42 @@ where
         Ok(MaybePendingTransactionReceipt::Receipt(receipt))
     }
 
+    fn get_transaction_execution_info(
+        &self,
+        parent_substrate_block_hash: B::Hash,
+        block_transactions: &BlockTransactions,
+        transaction_hash: TransactionHash,
+    ) -> Result<TransactionExecutionInfo, StarknetRpcApiError>
+    where
+        B: BlockT,
+    {
+        let (transactions_before, transaction_to_trace) =
+            split_block_tx_for_reexecution(block_transactions, transaction_hash).map_err(|e| {
+                log::error!("Failed to split block transactions for re-execution: {e}");
+                StarknetRpcApiError::InternalServerError
+            })?;
+
+        if transaction_to_trace.is_empty() {
+            return Err(StarknetRpcApiError::TxnHashNotFound);
+        }
+
+        if transaction_to_trace.len() > 1 {
+            log::error!("More than one transaction with the same transaction hash {:#?}", transaction_to_trace);
+            return Err(StarknetRpcApiError::InternalServerError);
+        }
+
+        let mut trace = self
+            .re_execute_transactions(parent_substrate_block_hash, transactions_before, transaction_to_trace, false)
+            .map_err(|e| {
+                log::error!("Failed to re-execute transactions: {e}");
+                StarknetRpcApiError::InternalServerError
+            })?;
+
+        let execution_info = trace.remove(0);
+
+        Ok(execution_info.0)
+    }
+
     fn get_events_for_tx_by_hash(
         &self,
         substrate_block_hash: B::Hash,
@@ -1607,11 +1638,10 @@ where
         let messages_sent = Vec::new();
         let events = Vec::new();
 
-        // TODO -- should Tx be simulated with `skip_validate`?
-        let skip_validate = true;
-        let skip_fee_charge = self.is_transaction_fee_disabled(self.get_best_block_hash())?;
+        let parent_substrate_block_hash = self.get_best_block_hash();
+        let pending_txs = self.get_pending_txs(parent_substrate_block_hash)?;
         let simulation =
-            self.simulate_tx(self.get_best_block_hash(), pending_tx.clone(), skip_validate, skip_fee_charge)?;
+            self.get_transaction_execution_info(parent_substrate_block_hash, &pending_txs, transaction_hash)?;
         let actual_fee =
             FeePayment { amount: Felt252Wrapper::from(simulation.actual_fee.0).into(), unit: PriceUnit::Wei };
         let execution_result = revert_error_to_execution_result(simulation.revert_error);
@@ -1674,20 +1704,6 @@ where
         };
 
         Ok(MaybePendingTransactionReceipt::PendingReceipt(receipt))
-    }
-
-    fn convert_error<T>(
-        &self,
-        best_block_hash: <B as BlockT>::Hash,
-        call_result: Result<T, DispatchError>,
-    ) -> Result<T, StarknetRpcApiError> {
-        match call_result {
-            Ok(val) => Ok(val),
-            Err(e) => {
-                let starknet_error = self.convert_dispatch_error(best_block_hash, e)?;
-                Err(starknet_error.into())
-            }
-        }
     }
 }
 
@@ -1767,13 +1783,12 @@ fn actual_resources_to_execution_resources(resources: ResourcesMapping) -> Execu
 }
 
 fn split_block_tx_for_reexecution(
-    block: &mp_block::Block,
+    block_transactions: &BlockTransactions,
     transaction_hash: TransactionHash,
 ) -> RpcResult<(
     Vec<blockifier::transaction::transaction_execution::Transaction>,
     Vec<blockifier::transaction::transaction_execution::Transaction>,
 )> {
-    let block_transactions = block.transactions();
     let tx_to_trace_idx = block_transactions
         .iter()
         .rposition(|tx| get_transaction_hash(tx) == &transaction_hash)