+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Next release
 
+- fix: starknet_call errs if contract nonexistent
 - fix: txn hash calculation and refactor
+- chore: remove all da/settlement related code
+- fix: re-execute txs instead of simulating for txn receipts
 - chore: rebase on latest blockifier
 - refactoring : Removed Redundant logs in madara
 - fix: transaction receipt fails for txs in the middle of a block
@@ -34,9 +37,13 @@
 - chore(db): changed the way hashes are encoded
 - refacto: reusable Eth client config for settlement/DA/other tasks
 - ci: add gomu gomu no gatling perfomrance test
+- feat(runtime): moved StarkEvents from Substrate events to runtime storage
+- feat(node): validate declare txs against local Sierra class in block import
+  queue
 - feat(runtime): moved StarkEvents from Substrate
 - feat(rpc/trace_api): add `trace_transaction`
 - fix(docker): fix dockerfile for `madara-node`
+- feat: Remove generic hasher from block hash computation
 
 ## v0.7.0